Перенесет функции для рендеринга из `utils.js` в `utils/render.js`, переименует utils.js в common.js
