+++ src/utils/common.js
@@ -1,10 +1,5 @@
 const MINUTES_PER_HOUR = 60;
 
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 export const getRandomArrayItem = (array) => {
   const randomIndex = getRandomIntegerNumber(0, array.length);
 
@@ -13,24 +8,6 @@ export const getRandomArrayItem = (array) => {
 
 export const getRandomIntegerNumber = (min, max) => {
   return min + Math.floor(max * Math.random());
-};
-
-export const createElement = (template) => {
-  const element = document.createElement(`div`);
-  element.innerHTML = template;
-
-  return element.firstChild;
-};
-
-export const renderElement = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
 };
 
 export const formatDate = (date, isLong) => {
